ref(metrics-extraction): Rename api hook file
